requirements.txt
torch
torchvision
torchaudio
opencv-python
pillow
scikit-learn
albumentations
pytesseract
fastapi
uvicorn
pandas